backend/requirements.txt
requests>=2.32.5
trafilatura>=2.0.0
qdrant-client>=1.16.1
cohere>=5.20.0
xmltodict>=1.0.2
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0